Move settings into admin.settings
