+++ app/views/ViewHelpers.scala
@@ -1,10 +1,11 @@
 package views
 
+import admin.settings.AllSettings
 import play.api.mvc.RequestHeader
 import play.twirl.api.{Html, HtmlFormat}
-import admin.AllSettings
-import admin.AllSettings.allSettingsCodec
 import io.circe.{Encoder, Printer}
+import admin.settings.AllSettings.allSettingsCodec
+import io.circe.Printer
 import io.circe.syntax._
 
 object ViewHelpers {
