+++ app/views/digitalSubscription.scala.html
@@ -1,4 +1,4 @@
-@import admin.AllSettings
+@import admin.settings.AllSettings
 @import assets.AssetsResolver
 @import com.gu.i18n.Currency.AUD
 @import com.gu.identity.play.IdUser
