+++ app/views/main.scala.html
@@ -1,6 +1,6 @@
+@import admin.settings.AllSettings
 @import assets.AssetsResolver
 @import views.ViewHelpers._
-@import admin.AllSettings
 
 @(
 	title: String,
