+++ app/views/oneOffContributions.scala.html
@@ -1,9 +1,9 @@
+@import admin.settings.AllSettings
 @import assets.AssetsResolver
 @import com.gu.identity.play.IdUser
 @import com.gu.support.config.StripeConfig
 @import helper.CSRF
 @import com.gu.i18n.Currency.AUD
-@import admin.AllSettings
 @(
         title: String,
         id: String,
