+++ app/views/optimizeScript.scala.html
@@ -1,6 +1,6 @@
 @import assets.AssetsResolver
-@import admin.AllSettings
-@import admin.SwitchState
+@import admin.settings.SwitchState
+@import admin.settings.AllSettings
 @()(implicit assets: AssetsResolver, settings: AllSettings)
 @if(settings.switches.optimize == SwitchState.On) {
     <!--- Optimize scripts, see: https://support.google.com/optimize/answer/7359264 -->
