+++ app/views/settingsScript.scala.html
@@ -1,5 +1,5 @@
-@import admin.AllSettings
 @import views.ViewHelpers.outputJson
+@import admin.settings.AllSettings
 
 @(settings: AllSettings)
